Some mouse tests can't run in parallel because MouseGrabView
